requirements.txt
torch>=2.2.0
torchvision
numpy
Pillow
boto3
sagemaker
supervision
opencv-python

